Adding BL-003 to Main

+++ creditcard_module/src/main/java/capstoneds2/creditcard_module/Model/Fatura.java
@@ -18,7 +18,7 @@ import java.time.LocalDate;
 public class Fatura {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer fatura_id;
+    private Long fatura_id;
 
     @ManyToOne
     @JoinColumn(name = "cartao_id", nullable = false)
